meta anywhere
I noticed that metadata extraction wasn't working with YouTube. This was
because unshrtn was expecting meta to be in head, but YouTube put it in
their body. So the query was relaxed to look for meta anywhere.

Also the test involving WashingtonPost no longer works so I switched it
over to look at RicRoll instead.